Trim CrewAI framework helper to focus on RAG pattern.
Consolidate the RAG layer into a single module, remove the eval harness
and duplicate bootstrap scripts, and shorten the README with replication notes.
