+++ 3_crewai/community_contributions/nvavrock/crewai_framework_helper/.env.example
@@ -1,14 +1,7 @@
 OPENAI_API_KEY=
 
-# Optional: override default question when running the crew
+# Optional
 # QUESTION=How do I use CrewBase with YAML agent configs?
-
-# Upstream CrewAI framework source (for RAG index)
-UPSTREAM_REPO_URL=https://github.com/crewAIInc/crewAI
-UPSTREAM_DIR=upstream/crewai
-UPSTREAM_REF=main
-
-# Index settings
-TARGET_REPO=upstream/crewai
-INDEX_PATHS=lib/crewai,lib/crewai-core,lib/crewai-tools,lib/crewai-files,lib/cli,docs/edge/en,AGENTS.md,README.md
-EMBEDDING_MODEL=text-embedding-3-small
+# UPSTREAM_REPO_URL=https://github.com/crewAIInc/crewAI
+# UPSTREAM_DIR=upstream/crewai
+# INDEX_PATHS=lib/crewai,lib/cli,docs/edge/en
