+++ 3_crewai/community_contributions/nvavrock/crewai_framework_helper/.gitignore
@@ -7,4 +7,3 @@ upstream/
 output/
 .uv/
 .DS_Store
-eval/results/
