+++ 3_crewai/community_contributions/nvavrock/crewai_framework_helper/README.md
@@ -1,22 +1,8 @@
 # CrewAI Framework Helper
 
-A **community contribution** for CrewAI week: a small crew that answers questions about the CrewAI framework using **RAG over indexed upstream source and docs** — not guesses.
+A small crew that answers CrewAI framework questions using **RAG over indexed upstream source and docs** — not guesses.
 
-Two agents run in sequence:
-
-1. **Framework researcher** — searches the index with `framework_search`
-2. **Framework advisor** — writes a cited answer to `output/answer.md`
-
-Related work: [Smart Coder](https://github.com/nvavrock/crewAI) uses the same RAG pattern to scaffold full projects. This contribution is a trimmed, course-friendly Q&A crew.
-
-## Requirements
-
-- Python `>=3.10,<3.14`
-- [uv](https://docs.astral.sh/uv/) package manager
-- **OpenAI API key** — required for bootstrap embeddings and agent LLMs (`gpt-4o-mini`)
-- **CrewAI CLI** (optional) — `uv tool install crewai==1.14.4` per course week 3; or use `uv run crewai run` below
-
-**Windows:** Chroma may require [MS Build Tools](https://github.com/bycloudai/InstallVSBuildToolsWindows) if install fails.
+**Pattern:** clone `crewAIInc/crewAI` → embed lib + docs into Chroma → expose a `framework_search` tool → two agents research then write a cited answer to `output/answer.md`.
 
 ## Setup
 
@@ -24,51 +10,28 @@ Related work: [Smart Coder](https://github.com/nvavrock/crewAI) uses the same RA
 cd 3_crewai/community_contributions/nvavrock/crewai_framework_helper
 cp .env.example .env   # add OPENAI_API_KEY
 uv sync
-```
-
-## Bootstrap the index (required before first run)
-
-Clones `crewAIInc/crewAI` into `upstream/crewai/` and embeds lib + docs into `.rag_index/`:
-
-```bash
-uv run bootstrap-index --yes
-```
-
-Or: `uv run python scripts/bootstrap_index.py --yes`
-
-## Run the crew
-
-```bash
+uv run bootstrap-index --yes   # clones upstream/crewai, builds .rag_index/
 uv run crewai run
 ```
 
-Or with a custom question:
+Custom question: `QUESTION="How do Flow @start and @listen work?" uv run crewai run`
+
+## What's unique here
+
+| Piece | File |
+|-------|------|
+| RAG (clone, index, search) | `src/crewai_framework_helper/rag.py` |
+| Custom tool | `src/crewai_framework_helper/tools/framework_search.py` |
+| Crew | `src/crewai_framework_helper/crew.py` + YAML configs |
+
+Everything else follows the standard `crewai create crew` layout. To replicate for another repo, copy `rag.py` and `framework_search.py`, point `UPSTREAM_REPO_URL` at your target, and adjust `INDEX_PATHS`.
+
+## Verify retrieval (optional)
 
 ```bash
-QUESTION="How do I use Flow with @start and @listen?" uv run crewai run
+uv run python -c "from crewai_framework_helper.rag import format_results, search; print(format_results(search('CrewBase decorator')))"
 ```
 
-Output: `output/answer.md` with explanation, code example, and **Sources** section.
+## Gitignored (local only)
 
-**Sources paths** are relative to `upstream/crewai/` (the local clone created during bootstrap). After each run, a note is appended to `answer.md` explaining that those files may not exist on your machine until you bootstrap, or if your upstream ref differs. The same paths exist on [crewAIInc/crewAI](https://github.com/crewAIInc/crewAI) on GitHub.
-
-## Example questions
-
-- How do I use CrewBase with YAML agent configs?
-- How do I add a custom tool with BaseTool?
-- How do I configure crew-level knowledge from text files?
-- What is the difference between @start and @listen on a Flow?
-
-## Verify the index (optional)
-
-```bash
-uv run eval-rag
-```
-
-Runs golden queries from `eval/rag_queries.yaml` (embedding search only during eval).
-
-## Local-only paths (gitignored)
-
-- `.rag_index/` — Chroma embeddings
-- `upstream/crewai/` — cloned framework repo
-- `output/` — generated answers
+`.rag_index/`, `upstream/crewai/`, `output/`
