--- 3_crewai/community_contributions/nvavrock/crewai_framework_helper/scripts/bootstrap_index.py
@@ -1,6 +1,0 @@
-"""One-shot: sync upstream CrewAI and build the RAG index."""
-
-from rag.bootstrap import main
-
-if __name__ == "__main__":
-    main()
